requirements.txt
discord.py
yt-dlp
python-dotenv
PyNaCl
ffmpeg-python